@@ -12,28 +12,37 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.1.2",
-    "@angular/common": "^5.1.2",
-    "@angular/compiler": "^5.1.2",
-    "@angular/core": "^5.1.2",
-    "@angular/forms": "^5.1.2",
-    "@angular/http": "^5.1.2",
-    "@angular/platform-browser": "^5.1.2",
-    "@angular/platform-browser-dynamic": "^5.1.2",
-    "@angular/router": "^5.1.2",
-    "core-js": "^2.4.1",
+    "@angular-devkit/core": "^0.2.0",
+    "@angular-devkit/schematics": "^0.2.0",
+    "@angular/animations": "^5.2.3",
+    "@angular/cdk": "^5.1.1",
+    "@angular/common": "^5.2.3",
+    "@angular/compiler": "^5.2.3",
+    "@angular/core": "^5.2.3",
+    "@angular/forms": "^5.2.3",
+    "@angular/http": "^5.2.3",
+    "@angular/material": "^5.1.1",
+    "@angular/platform-browser": "^5.2.3",
+    "@angular/platform-browser-dynamic": "^5.2.3",
+    "@angular/router": "^5.2.3",
+    "@schematics/angular": "^0.2.0",
+    "@schematics/schematics": "^0.2.0",
+    "core-js": "^2.5.3",
     "npm": "^5.5.1",
-    "rxjs": "^5.5.2",
-    "zone.js": "^0.8.14"
+    "rxjs": "^5.5.6",
+    "zone.js": "^0.8.20",
+    "hammerjs": "^2.0.8",
+    "web-animations-js": "^2.3.1",
+    "angular-in-memory-web-api": "0.5.3"
   },
   "devDependencies": {
     "@angular/cli": "1.6.3",
-    "@angular/compiler-cli": "^5.1.2",
-    "@angular/language-service": "^5.1.2",
+    "@angular/compiler-cli": "^5.2.3",
+    "@angular/language-service": "^5.2.3",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.5.2",
-    "angular-ide": "^0.9.38",
+    "angular-ide": "^0.9.39",
     "codelyzer": "~4.0.2",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
