@@ -1,11 +1,12 @@
-import { ModuleWithProviders, NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
-import { CommonModule } from '@angular/common';
+import {ModuleWithProviders, NgModule} from '@angular/core';
+import {RouterModule} from '@angular/router';
+import {CommonModule} from '@angular/common';
+import {MatCardModule, MatDividerModule, MatListModule} from '@angular/material';
 
-import { SubmitterListComponent } from './submitter-list.component';
-import { SubmitterListItemComponent } from './submitter-list-item.component';
-import { SubmitterListResolver } from './submitter-list-resolver.service';
-import { SharedModule } from '../shared';
+import {SubmitterListComponent} from './submitter-list.component';
+import {SubmitterListItemComponent} from './submitter-list-item.component';
+import {SubmitterListResolver} from './submitter-list-resolver.service';
+import {SharedModule} from '../shared';
 
 const submitterRouting: ModuleWithProviders = RouterModule.forChild([
   {
@@ -21,7 +22,10 @@ const submitterRouting: ModuleWithProviders = RouterModule.forChild([
   imports: [
     submitterRouting,
     SharedModule,
-    CommonModule
+    CommonModule,
+    MatCardModule,
+    MatDividerModule,
+    MatListModule,
   ],
   declarations: [
     SubmitterListComponent,
